requirements.txt
Pillow==10.4.0
python-dotenv==1.0.1
aiogram~=3.20.0.post0
aio-pika~=9.5.5